added test file system storage

+++ src/Sa.HybridFileStorage.FileSystem/FileSystemStorage.cs
@@ -5,7 +5,7 @@ namespace Sa.HybridFileStorage.FileSystem;
 
 internal class FileSystemStorage(FileSystemStorageOptions options, ICurrentTimeProvider currentTime) : IFileStorage
 {
-    private readonly string _basePath = options.BasePath;
+    private readonly string _basePath = Path.TrimEndingDirectorySeparator(options.BasePath);
 
     public string StorageType => options.StorageType ?? "file";
 
@@ -13,24 +13,28 @@ internal class FileSystemStorage(FileSystemStorageOptions options, ICurrentTimeP
 
     public async Task<StorageResult> UploadFileAsync(UploadFileInput metadata, Stream fileStream, CancellationToken cancellationToken)
     {
-        string filePath = Path.Combine(_basePath, metadata.TenantId.ToString(), metadata.FileName).Replace('\\', '/');
-
-        string? dir = Path.GetDirectoryName(filePath);
-        if (!Directory.Exists(dir))
+        if (IsReadOnly)
         {
-            Directory.CreateDirectory(dir!);
+            throw new InvalidOperationException("Cannot upload file. All storage options are read-only.");
         }
 
-        var fileId = new Uri(filePath).AbsolutePath;
+        string filePath = $"{_basePath}/{metadata.TenantId}/{metadata.FileName.Replace('\\', '/')}";
+
+        string? dir = Path.GetDirectoryName(filePath);
+        EnsureDirectory(dir);
 
         using var fileStreamOutput = new FileStream(filePath, FileMode.Create, FileAccess.Write);
         await fileStream.CopyToAsync(fileStreamOutput, cancellationToken);
-        return new StorageResult(fileId, filePath, StorageType, currentTime.GetUtcNow());
+
+        var fileId = FilePathToId(filePath);
+        var fileAbsolute = Path.GetFullPath(filePath);
+
+        return new StorageResult(fileId, fileAbsolute, StorageType, currentTime.GetUtcNow());
     }
 
     public async Task<bool> DownloadFileAsync(string fileId, Func<Stream, CancellationToken, Task> loadStream, CancellationToken cancellationToken)
     {
-        var filePath = Path.Combine(_basePath, fileId);
+        var filePath = FileIdToPath(fileId);
         if (File.Exists(filePath))
         {
             using var fs = new FileStream(filePath, FileMode.Open, FileAccess.Read);
@@ -42,7 +46,12 @@ internal class FileSystemStorage(FileSystemStorageOptions options, ICurrentTimeP
 
     public Task<bool> DeleteFileAsync(string fileId, CancellationToken cancellationToken)
     {
-        var filePath = Path.Combine(_basePath, fileId);
+        if (IsReadOnly)
+        {
+            throw new InvalidOperationException("Cannot delete file. All storage options are read-only.");
+        }
+
+        var filePath = FileIdToPath(fileId);
         if (File.Exists(filePath))
         {
             File.Delete(filePath);
@@ -51,5 +60,36 @@ internal class FileSystemStorage(FileSystemStorageOptions options, ICurrentTimeP
         return Task.FromResult(false);
     }
 
-    public bool CanProcessFileId(string fileId) => fileId.StartsWith(StorageType);
+    public bool CanProcessFileId(string fileId) 
+        => fileId.StartsWith(FilePathToId(_basePath));
+
+    private static void EnsureDirectory(string? dir)
+    {
+        if (!Directory.Exists(dir))
+        {
+            Directory.CreateDirectory(dir!);
+        }
+    }
+
+    private string FilePathToId(string filePath) => $"{StorageType}://{filePath}";
+
+    private static string FileIdToPath(string fileId)
+    {
+        if (string.IsNullOrWhiteSpace(fileId))
+        {
+            throw new ArgumentException("File ID cannot be null or empty.", nameof(fileId));
+        }
+
+        ReadOnlySpan<char> span = fileId.AsSpan();
+
+        int separatorIndex = span.IndexOf("://");
+        if (separatorIndex == -1)
+        {
+            throw new FormatException("Invalid file ID format.");
+        }
+
+        ReadOnlySpan<char> filePath = span[(separatorIndex + 3)..]; // +3 for skip "://"
+
+        return filePath.ToString();
+    }
 }
